// admin-dashboard/store/film/action-type.ts
export enum ActionTypes {
  GET_FILMS = "GET_FILMS",
  GET_FILM = "GET_FILM",
  UPDATE_FILM = "UPDATE_FILM",
  PUBLISH_FILM = "PUBLISH_FILM",
  HARD_DELETE_FILM = "HARD_DELETE_FILM",
  DELETE_FILM = "DELETE_FILM",
  UPDATE_FILM_THUMBNAIL = "UPDATE_FILM_THUMBNAIL",
  UPLOAD_FILM_TRAILER = "UPLOAD_FILM_TRAILER",
  CREATE_FILM = "CREATE_FILM",
  UPLOAD_FILM = "UPLOAD_FILM",
}